requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asttokens 
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama 
comm
debugpy 
decorator
distlib==0.3.8
exceptiongroup 
executing 
filelock==3.16.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata 
ipykernel 
ipython 
jedi 
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client 
jupyter_core 
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-google-genai==2.0.1
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.132
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib-inline 
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.9.2
nest_asyncio 
numpy==1.26.4
orjson==3.10.7
packaging 
pandas==2.2.3
parso 
pickleshare 
pillow==10.4.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs 
prompt_toolkit 
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil 
pure_eval 
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments 
pyparsing==3.1.4
pypdf==5.0.1
PyPDF2==3.0.1
python-dateutil
python-dotenv==1.0.1
pytz==2024.2
pywin32==305.1
PyYAML==6.0.2
pyzmq
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
setuptools==75.1.0
six
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
stack-data
streamlit==1.39.0
tenacity==8.5.0
toml==0.10.2
tornado 
tqdm==4.66.5
traitlets 
typing-inspect==0.9.0
typing_extensions 
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.4
watchdog==5.0.3
wcwidth 
wheel==0.44.0
yarl==1.14.0
zipp 